Moved ViewModelFactory to factories directive
